requirements.txt
fastapi===0.115.12
uvicorn===0.34.3
pydantic===2.11.5
pyinstaller===6.14.1
python-escpos===3.1
pyusb===1.3.1
jinja2===3.0.3
python-multipart===0.0.20
zeroconf===0.147.0
